Split mealplan tests into separate files for models, views, and forms.
