+++ mealplan/tests/test_models.py
@@ -1,0 +1,6 @@
+"""Tests Mealplan model and its associated form."""
+
+from django.test import TestCase
+
+class MealplanModelTest(TestCase):
+
